Fix dropped argument in HashingUtils 5-arg Object hash
The 5-arg Object overload was forwarding only obj0..obj3 to the int
overload, silently dropping obj4. Also align LongHashingUtils.hash 3-arg
signature with its 2/4/5-arg siblings (int parameters) and strengthen
the 5-arg HashingUtilsTest to detect the missing-arg regression.

+++ internal-api/src/main/java/datadog/trace/util/LongHashingUtils.java
@@ -60,7 +60,7 @@ public final class LongHashingUtils {
     return hash(intHash(obj0), intHash(obj1), intHash(obj2));
   }
 
-  public static final long hash(long hash0, long hash1, long hash2) {
+  public static final long hash(int hash0, int hash1, int hash2) {
     // DQH - Micro-optimizing, 31L * 31L will constant fold
     // Since there are multiple execution ports for load & store,
     // this will make good use of the core.
